refactor: move tag files to another folder
